Changing scrollview to component
