+++ src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './header/header';
 export * from './pollTemplatesGrid/pollTemplatesGrid';
 export * from './pollTemplate/pollTemplate';
+export * from './scrollView/scrollView';
 // export * from './footer';
 // export * from '../albums/albumListItem';
